backend/requirements.txt
fastapi
uvicorn

langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface

faiss-cpu

sentence-transformers
transformers
huggingface-hub

pypdf

python-dotenv
